--- packages/components/src/components/Navigation/components/NavigationItem/NavigationCollectionItem.tsx
@@ -1,10 +1,0 @@
-import { ComponentProps, ComponentType } from "react";
-import { Item, ItemProps } from "react-stately";
-
-export interface NavigationCollectionItemProps<T = never> extends ItemProps<T> {
-  isCurrent?: boolean;
-  linkComponent?: ComponentType<Omit<ComponentProps<"a">, "ref">>;
-}
-
-export const NavigationCollectionItem =
-  Item as ComponentType<NavigationCollectionItemProps>;
