+++ packages/components/src/components/Navigation/components/NavigationItem/NavigationItem.module.scss
@@ -1,31 +1,29 @@
 .navigationItem {
-  a {
-    display: flex;
-    flex-direction: row;
+  display: flex;
+  flex-direction: row;
 
-    border: none;
-    border-radius: var(--navigation-item--corner-radius);
-    background-color: var(--navigation-item--background-color--default);
-    padding: var(--navigation-item--padding-y) var(--navigation-item--padding-x);
+  border: none;
+  border-radius: var(--navigation-item--corner-radius);
+  background-color: var(--navigation-item--background-color--default);
+  padding: var(--navigation-item--padding-y) var(--navigation-item--padding-x);
 
-    cursor: default;
+  cursor: default;
 
-    &[href] {
-      cursor: pointer;
-    }
+  outline-color: var(--navigation-item--focus-outline-color);
+  outline-offset: var(--navigation-item--focus-outline-offset);
+  outline-width: var(--navigation-item--focus-outline-width);
 
-    &:focus {
-      background-color: var(--navigation-item--background-color--hover);
-    }
+  &:hover {
+    background-color: var(--navigation-item--background-color--hover);
+  }
 
-    &[aria-current="page"] {
-      background-color: var(--navigation-item--background-color--current);
-      font-weight: var(--navigation-item--font-weight--current);
-      color: var(--navigation-item--color--current);
+  &[aria-current="page"] {
+    background-color: var(--navigation-item--background-color--current);
+    font-weight: var(--navigation-item--font-weight--current);
+    color: var(--navigation-item--color--current);
 
-      .icon {
-        color: var(--navigation-item--icon-color--current);
-      }
+    .icon {
+      color: var(--navigation-item--icon-color--current);
     }
   }
 }
