+++ packages/design-tokens/src/components/navigation.yml
@@ -1,6 +1,6 @@
 navigation:
   item-to-item-spacing:
-    value: "{size-px.xs}"
+    value: "{size-px.s}"
 
 navigation-item:
   font-weight:
@@ -11,9 +11,16 @@ navigation-item:
   padding-x:
     value: "{size-px.s}"
   padding-y:
-    value: "{size-px.s}"
+    value: "{size-px.xs}"
   icon-to-text-spacing:
     value: 0.5em
+  focus-outline-color:
+    value: "{focus-outline-color}"
+  focus-outline-offset:
+    value: "{size-px.xxs}"
+  focus-outline-width:
+    value: "{border-width.200}"
+
   color:
     current:
       value: "{info.color.800}"
