+++ packages/docs/src/app/_components/layout/MainNavigation/components/GroupText.tsx
@@ -5,7 +5,7 @@ interface Props {
   children: string;
 }
 
-export const GroupHeadingText: FC<Props> = (props) => {
+export const GroupText: FC<Props> = (props) => {
   const { children } = props;
 
   return <>{humanizeString(children.replaceAll(/^[0-9]+/g, ""))}</>;
